Support for addition of SUT & Emissary strategies
Split schema and validator from single Job into BrowserApp and Api to
support addition of SUT and Emissary strategies in app-tester.
Fix logic that sends Job to back-end to include modifications
(sanitisation) that the schema makes.
Add example Job files.
Change CUI text Bugs to Alerts.
Provide clarification in README on back-end response messages.
