+++ config/config.test.json
@@ -9,7 +9,7 @@
     "host": "127.0.0.1"
   },
   "job": {
-    "fileUri": "./testResources/jobs/job_1.0.0-alpha.3_local"
+    "fileUri": "./testResources/jobs/job_2.0.0-alpha.3_local"
   },
   "modulePaths": {
     "blessed": "../../testResources/blessedFake"
