+++ test/presenter/apiDecoratingAdapter.js
@@ -37,10 +37,10 @@ const Model = require('../../src/models/model');
 const cUiPath = '../../src/view/cUi';
 const apiDecoratingAdapterPath = '../../src/presenter/apiDecoratingAdapter';
 
-// As stored in the `request` object body from file: /testResources/jobs/job_1.0.0-alpha.3
-const expectedJob = '\"{\\n  \\\"data\\\": {\\n    \\\"type\\\": \\\"job\\\",\\n    \\\"attributes\\\": {\\n      \\\"version\\\": \\\"1.0.0-alpha.3\\\",\\n      \\\"sutAuthentication\\\": {\\n        \\\"route\\\": \\\"/login\\\",\\n        \\\"usernameFieldLocater\\\": \\\"userName\\\",\\n        \\\"passwordFieldLocater\\\": \\\"password\\\",\\n        \\\"submit\\\": \\\"btn btn-danger\\\",\\n        \\\"expectedPageSourceSuccess\\\": \\\"Log Out\\\"\\n      },\\n      \\\"sutIp\\\": \\\"pt-sut-cont\\\",\\n      \\\"sutPort\\\": 4000,\\n      \\\"sutProtocol\\\": \\\"http\\\",\\n      \\\"browser\\\": \\\"chrome\\\",\\n      \\\"loggedInIndicator\\\": \\\"<p>Found. Redirecting to <a href=\\\\\\\"\\\\/dashboard\\\\\\\">\\\\/dashboard<\\\\/a><\\\\/p>\\\"\\n    },\\n    \\\"relationships\\\": {\\n      \\\"data\\\": [{\\n        \\\"type\\\": \\\"tlsScanner\\\",\\n        \\\"id\\\": \\\"NA\\\"\\n      },\\n      {\\n        \\\"type\\\": \\\"appScanner\\\",\\n        \\\"id\\\": \\\"lowPrivUser\\\"\\n      },\\n      {\\n        \\\"type\\\": \\\"appScanner\\\",\\n        \\\"id\\\": \\\"adminUser\\\"\\n      }]\\n    }\\n  },\\n  \\\"included\\\": [\\n    {\\n      \\\"type\\\": \\\"tlsScanner\\\",\\n      \\\"id\\\": \\\"NA\\\",\\n      \\\"attributes\\\": {\\n        \\\"tlsScannerSeverity\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 3\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"lowPrivUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"username\\\": \\\"user1\\\",\\n        \\\"password\\\": \\\"User1_123\\\",\\n        \\\"aScannerAttackStrength\\\": \\\"HIGH\\\",\\n        \\\"aScannerAlertThreshold\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 12\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [{\\n          \\\"type\\\": \\\"route\\\",\\n          \\\"id\\\": \\\"/profile\\\"\\n        }]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"adminUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"username\\\": \\\"admin\\\",\\n        \\\"password\\\": \\\"Admin_123\\\"\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [{\\n          \\\"type\\\": \\\"route\\\",\\n          \\\"id\\\": \\\"/memos\\\"\\n        },\\n        {\\n          \\\"type\\\": \\\"route\\\",\\n          \\\"id\\\": \\\"/profile\\\"\\n        }]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/profile\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\\"name\\\": \\\"firstName\\\", \\\"value\\\": \\\"PurpleJohn\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"lastName\\\", \\\"value\\\": \\\"PurpleDoe\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"ssn\\\", \\\"value\\\": \\\"PurpleSSN\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"dob\\\", \\\"value\\\": \\\"12235678\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"bankAcc\\\", \\\"value\\\": \\\"PurpleBankAcc\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"bankRouting\\\", \\\"value\\\": \\\"0198212#\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"address\\\", \\\"value\\\": \\\"PurpleAddress\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"website\\\", \\\"value\\\": \\\"https://purpleteam-labs.com\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"_csrf\\\", \\\"value\\\": \\\"\\\"},\\n          {\\\"name\\\": \\\"submit\\\", \\\"value\\\": \\\"\\\"}\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"submit\\\"\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/memos\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\\"name\\\": \\\"memo\\\", \\\"value\\\": \\\"PurpleMemo\\\", \\\"visible\\\": true}\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"btn btn-primary\\\"\\n      }\\n    }\\n  ]\\n}\\n\"'; // eslint-disable-line no-useless-escape
-// As stored in the `request` object body from file: /testResources/jobs/job_1.0.0-alpha.3_missing_type_of_appScanner
-const expectedJobMissingTypeAppScanner = '\"{\\n  \\\"data\\\": {\\n    \\\"type\\\": \\\"job\\\",\\n    \\\"attributes\\\": {\\n      \\\"version\\\": \\\"1.0.0-alpha.3\\\",\\n      \\\"sutAuthentication\\\": {\\n        \\\"route\\\": \\\"/login\\\",\\n        \\\"usernameFieldLocater\\\": \\\"userName\\\",\\n        \\\"passwordFieldLocater\\\": \\\"password\\\",\\n        \\\"submit\\\": \\\"btn btn-danger\\\",\\n        \\\"expectedPageSourceSuccess\\\": \\\"Log Out\\\"\\n      },\\n      \\\"sutIp\\\": \\\"pt-sut-cont\\\",\\n      \\\"sutPort\\\": 4000,\\n      \\\"sutProtocol\\\": \\\"http\\\",\\n      \\\"browser\\\": \\\"chrome\\\",\\n      \\\"loggedInIndicator\\\": \\\"<p>Found. Redirecting to <a href=\\\\\\\"\\\\/dashboard\\\\\\\">\\\\/dashboard<\\\\/a><\\\\/p>\\\"\\n    },\\n    \\\"relationships\\\": {\\n      \\\"data\\\": [{\\n        \\\"type\\\": \\\"appScanner\\\",\\n        \\\"id\\\": \\\"lowPrivUser\\\"\\n      },\\n      {\\n        \\\"type\\\": \\\"appScanner\\\",\\n        \\\"id\\\": \\\"adminUser\\\"\\n      }]\\n    }\\n  },\\n  \\\"included\\\": [\\n    {\\n      \\\"type\\\": \\\"tlsScanner\\\",\\n      \\\"id\\\": \\\"NA\\\",\\n      \\\"attributes\\\": {\\n        \\\"tlsScannerSeverity\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 3\\n      }\\n    },\\n    {\\n      \\\"id\\\": \\\"lowPrivUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"username\\\": \\\"user1\\\",\\n        \\\"password\\\": \\\"User1_123\\\",\\n        \\\"aScannerAttackStrength\\\": \\\"HIGH\\\",\\n        \\\"aScannerAlertThreshold\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 12\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [{\\n          \\\"type\\\": \\\"route\\\",\\n          \\\"id\\\": \\\"/profile\\\"\\n        }]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"adminUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"username\\\": \\\"admin\\\",\\n        \\\"password\\\": \\\"Admin_123\\\"\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [{\\n          \\\"type\\\": \\\"route\\\",\\n          \\\"id\\\": \\\"/memos\\\"\\n        },\\n        {\\n          \\\"type\\\": \\\"route\\\",\\n          \\\"id\\\": \\\"/profile\\\"\\n        }]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/profile\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\\"name\\\": \\\"firstName\\\", \\\"value\\\": \\\"PurpleJohn\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"lastName\\\", \\\"value\\\": \\\"PurpleDoe\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"ssn\\\", \\\"value\\\": \\\"PurpleSSN\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"dob\\\", \\\"value\\\": \\\"12235678\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"bankAcc\\\", \\\"value\\\": \\\"PurpleBankAcc\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"bankRouting\\\", \\\"value\\\": \\\"0198212#\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"address\\\", \\\"value\\\": \\\"PurpleAddress\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"website\\\", \\\"value\\\": \\\"https://purpleteam-labs.com\\\", \\\"visible\\\": true},\\n          {\\\"name\\\": \\\"_csrf\\\", \\\"value\\\": \\\"\\\"},\\n          {\\\"name\\\": \\\"submit\\\", \\\"value\\\": \\\"\\\"}\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"submit\\\"\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/memos\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\\"name\\\": \\\"memo\\\", \\\"value\\\": \\\"PurpleMemo\\\", \\\"visible\\\": true}\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"btn btn-primary\\\"\\n      }\\n    }\\n  ]\\n}\\n\"'; // eslint-disable-line no-useless-escape
+// As stored in the `request` object body from file: /testResources/jobs/job_2.0.0-alpha.3
+const expectedJob = '\"{\\n  \\\"data\\\": {\\n    \\\"type\\\": \\\"BrowserApp\\\",\\n    \\\"attributes\\\": {\\n      \\\"version\\\": \\\"2.0.0-alpha.3\\\",\\n      \\\"sutAuthentication\\\": {\\n        \\\"sitesTreeSutAuthenticationPopulationStrategy\\\": \\\"FormStandard\\\",\\n        \\\"emissaryAuthenticationStrategy\\\": \\\"FormStandard\\\",\\n        \\\"route\\\": \\\"/login\\\",\\n        \\\"usernameFieldLocater\\\": \\\"userName\\\",\\n        \\\"passwordFieldLocater\\\": \\\"password\\\",\\n        \\\"submit\\\": \\\"btn btn-danger\\\",\\n        \\\"expectedPageSourceSuccess\\\": \\\"Log Out\\\"\\n      },\\n      \\\"sutIp\\\": \\\"pt-sut-cont\\\",\\n      \\\"sutPort\\\": 4000,\\n      \\\"sutProtocol\\\": \\\"http\\\",\\n      \\\"browser\\\": \\\"chrome\\\",\\n      \\\"loggedInIndicator\\\": \\\"<p>Found. Redirecting to <a href=\\\\\\\"/dashboard\\\\\\\">/dashboard</a></p>\\\"\\n    },\\n    \\\"relationships\\\": {\\n      \\\"data\\\": [\\n        {\\n          \\\"type\\\": \\\"tlsScanner\\\",\\n          \\\"id\\\": \\\"NA\\\"\\n        },\\n        {\\n          \\\"type\\\": \\\"appScanner\\\",\\n          \\\"id\\\": \\\"lowPrivUser\\\"\\n        },\\n        {\\n          \\\"type\\\": \\\"appScanner\\\",\\n          \\\"id\\\": \\\"adminUser\\\"\\n        }\\n      ]\\n    }\\n  },\\n  \\\"included\\\": [\\n    {\\n      \\\"type\\\": \\\"tlsScanner\\\",\\n      \\\"id\\\": \\\"NA\\\",\\n      \\\"attributes\\\": {\\n        \\\"tlsScannerSeverity\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 3\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"lowPrivUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"sitesTreePopulationStrategy\\\": \\\"WebDriverStandard\\\",\\n        \\\"spiderStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scannersStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scanningStrategy\\\": \\\"Standard\\\",\\n        \\\"postScanningStrategy\\\": \\\"Standard\\\",\\n        \\\"reportingStrategy\\\": \\\"Standard\\\",\\n        \\\"username\\\": \\\"user1\\\",\\n        \\\"password\\\": \\\"User1_123\\\",\\n        \\\"aScannerAttackStrength\\\": \\\"HIGH\\\",\\n        \\\"aScannerAlertThreshold\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 12\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [\\n          {\\n            \\\"type\\\": \\\"route\\\",\\n            \\\"id\\\": \\\"/profile\\\"\\n          }\\n        ]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"adminUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"sitesTreePopulationStrategy\\\": \\\"WebDriverStandard\\\",\\n        \\\"spiderStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scannersStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scanningStrategy\\\": \\\"Standard\\\",\\n        \\\"postScanningStrategy\\\": \\\"Standard\\\",\\n        \\\"reportingStrategy\\\": \\\"Standard\\\",\\n        \\\"username\\\": \\\"admin\\\",\\n        \\\"password\\\": \\\"Admin_123\\\"\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [\\n          {\\n            \\\"type\\\": \\\"route\\\",\\n            \\\"id\\\": \\\"/memos\\\"\\n          },\\n          {\\n            \\\"type\\\": \\\"route\\\",\\n            \\\"id\\\": \\\"/profile\\\"\\n          }\\n        ]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/profile\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\n            \\\"name\\\": \\\"firstName\\\",\\n            \\\"value\\\": \\\"PurpleJohn\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"lastName\\\",\\n            \\\"value\\\": \\\"PurpleDoe\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"ssn\\\",\\n            \\\"value\\\": \\\"PurpleSSN\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"dob\\\",\\n            \\\"value\\\": \\\"12235678\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"bankAcc\\\",\\n            \\\"value\\\": \\\"PurpleBankAcc\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"bankRouting\\\",\\n            \\\"value\\\": \\\"0198212#\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"address\\\",\\n            \\\"value\\\": \\\"PurpleAddress\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"website\\\",\\n            \\\"value\\\": \\\"https://purpleteam-labs.com\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"_csrf\\\",\\n            \\\"value\\\": \\\"\\\"\\n          },\\n          {\\n            \\\"name\\\": \\\"submit\\\",\\n            \\\"value\\\": \\\"\\\"\\n          }\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"submit\\\"\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/memos\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\n            \\\"name\\\": \\\"memo\\\",\\n            \\\"value\\\": \\\"PurpleMemo\\\",\\n            \\\"visible\\\": true\\n          }\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"btn btn-primary\\\"\\n      }\\n    }\\n  ]\\n}\"'; // eslint-disable-line no-useless-escape
+// As stored in the `request` object body from file: /testResources/jobs/job_2.0.0-alpha.3_missing_type_of_appScanner
+const expectedJobMissingTypeAppScanner = '\"{\\n  \\\"data\\\": {\\n    \\\"type\\\": \\\"BrowserApp\\\",\\n    \\\"attributes\\\": {\\n      \\\"version\\\": \\\"2.0.0-alpha.3\\\",\\n      \\\"sutAuthentication\\\": {\\n        \\\"sitesTreeSutAuthenticationPopulationStrategy\\\": \\\"FormStandard\\\",\\n        \\\"emissaryAuthenticationStrategy\\\": \\\"FormStandard\\\",\\n        \\\"route\\\": \\\"/login\\\",\\n        \\\"usernameFieldLocater\\\": \\\"userName\\\",\\n        \\\"passwordFieldLocater\\\": \\\"password\\\",\\n        \\\"submit\\\": \\\"btn btn-danger\\\",\\n        \\\"expectedPageSourceSuccess\\\": \\\"Log Out\\\"\\n      },\\n      \\\"sutIp\\\": \\\"pt-sut-cont\\\",\\n      \\\"sutPort\\\": 4000,\\n      \\\"sutProtocol\\\": \\\"http\\\",\\n      \\\"browser\\\": \\\"chrome\\\",\\n      \\\"loggedInIndicator\\\": \\\"<p>Found. Redirecting to <a href=\\\\\\\"/dashboard\\\\\\\">/dashboard</a></p>\\\"\\n    },\\n    \\\"relationships\\\": {\\n      \\\"data\\\": [\\n        {\\n          \\\"type\\\": \\\"tlsScanner\\\",\\n          \\\"id\\\": \\\"NA\\\"\\n        },\\n        {\\n          \\\"type\\\": \\\"appScanner\\\",\\n          \\\"id\\\": \\\"lowPrivUser\\\"\\n        },\\n        {\\n          \\\"id\\\": \\\"adminUser\\\"\\n        }\\n      ]\\n    }\\n  },\\n  \\\"included\\\": [\\n    {\\n      \\\"type\\\": \\\"tlsScanner\\\",\\n      \\\"id\\\": \\\"NA\\\",\\n      \\\"attributes\\\": {\\n        \\\"tlsScannerSeverity\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 3\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"lowPrivUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"sitesTreePopulationStrategy\\\": \\\"WebDriverStandard\\\",\\n        \\\"spiderStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scannersStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scanningStrategy\\\": \\\"Standard\\\",\\n        \\\"postScanningStrategy\\\": \\\"Standard\\\",\\n        \\\"reportingStrategy\\\": \\\"Standard\\\",\\n        \\\"username\\\": \\\"user1\\\",\\n        \\\"password\\\": \\\"User1_123\\\",\\n        \\\"aScannerAttackStrength\\\": \\\"HIGH\\\",\\n        \\\"aScannerAlertThreshold\\\": \\\"LOW\\\",\\n        \\\"alertThreshold\\\": 12\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [\\n          {\\n            \\\"type\\\": \\\"route\\\",\\n            \\\"id\\\": \\\"/profile\\\"\\n          }\\n        ]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"appScanner\\\",\\n      \\\"id\\\": \\\"adminUser\\\",\\n      \\\"attributes\\\": {\\n        \\\"sitesTreePopulationStrategy\\\": \\\"WebDriverStandard\\\",\\n        \\\"spiderStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scannersStrategy\\\": \\\"BrowserAppStandard\\\",\\n        \\\"scanningStrategy\\\": \\\"Standard\\\",\\n        \\\"postScanningStrategy\\\": \\\"Standard\\\",\\n        \\\"reportingStrategy\\\": \\\"Standard\\\",\\n        \\\"username\\\": \\\"admin\\\",\\n        \\\"password\\\": \\\"Admin_123\\\"\\n      },\\n      \\\"relationships\\\": {\\n        \\\"data\\\": [\\n          {\\n            \\\"type\\\": \\\"route\\\",\\n            \\\"id\\\": \\\"/memos\\\"\\n          },\\n          {\\n            \\\"type\\\": \\\"route\\\",\\n            \\\"id\\\": \\\"/profile\\\"\\n          }\\n        ]\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/profile\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\n            \\\"name\\\": \\\"firstName\\\",\\n            \\\"value\\\": \\\"PurpleJohn\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"lastName\\\",\\n            \\\"value\\\": \\\"PurpleDoe\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"ssn\\\",\\n            \\\"value\\\": \\\"PurpleSSN\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"dob\\\",\\n            \\\"value\\\": \\\"12235678\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"bankAcc\\\",\\n            \\\"value\\\": \\\"PurpleBankAcc\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"bankRouting\\\",\\n            \\\"value\\\": \\\"0198212#\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"address\\\",\\n            \\\"value\\\": \\\"PurpleAddress\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"website\\\",\\n            \\\"value\\\": \\\"https://purpleteam-labs.com\\\",\\n            \\\"visible\\\": true\\n          },\\n          {\\n            \\\"name\\\": \\\"_csrf\\\",\\n            \\\"value\\\": \\\"\\\"\\n          },\\n          {\\n            \\\"name\\\": \\\"submit\\\",\\n            \\\"value\\\": \\\"\\\"\\n          }\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"submit\\\"\\n      }\\n    },\\n    {\\n      \\\"type\\\": \\\"route\\\",\\n      \\\"id\\\": \\\"/memos\\\",\\n      \\\"attributes\\\": {\\n        \\\"attackFields\\\": [\\n          {\\n            \\\"name\\\": \\\"memo\\\",\\n            \\\"value\\\": \\\"PurpleMemo\\\",\\n            \\\"visible\\\": true\\n          }\\n        ],\\n        \\\"method\\\": \\\"POST\\\",\\n        \\\"submit\\\": \\\"btn btn-primary\\\"\\n      }\\n    }\\n  ]\\n}\"'; // eslint-disable-line no-useless-escape
 
 
 describe('apiDecoratingAdapter', () => {
@@ -58,21 +58,20 @@ describe('apiDecoratingAdapter', () => {
       const expectedArgPasssedToTestPlan = [{
         name: 'app',
         message: `@app_scan
-        Feature: Web application free of security vulnerabilities known to Zap
+        Feature: Web application free of security vulnerabilities known to the Emissary
         
         # Before hooks are run before Background
         
         Background:
           Given a new Test Session based on each Build User supplied appScanner resourceObject
-          And each Build User supplied route of each appScanner resourceObject is navigated
-          And a new scanning session based on each Build User supplied appScanner resourceObject
+          And the Emissary sites tree is populated with each Build User supplied route of each appScanner resourceObject
+          And the Emissary authentication is configured for the SUT
           And the application is spidered for each appScanner resourceObject
-          And all active scanners are disabled
         
-        Scenario: The application should not contain vulnerabilities known to Zap that exceed the Build User defined threshold
-          Given all active scanners are enabled 
+        Scenario: The application should not contain vulnerabilities known to the Emissary that exceed the Build User defined threshold
+          Given the active scanners are configured
           When the active scan is run
-          Then the vulnerability count should not exceed the Build User defined threshold of vulnerabilities known to Zap
+          Then the vulnerability count should not exceed the Build User defined threshold of vulnerabilities known to the Emissary
         
         
         
@@ -173,9 +172,9 @@ describe('apiDecoratingAdapter', () => {
 
     it('- on - invalid JSON syntax - should print useful error message', async (flags) => {
       const { context: { rewiredApi, critStub } } = flags;
-      const jobFileContents = await (async () => readFileAsync(`${process.cwd()}/testResources/jobs/job_1.0.0-alpha.3_local_missing_comma`, { encoding: 'utf8' }))();
+      const jobFileContents = await (async () => readFileAsync(`${process.cwd()}/testResources/jobs/job_2.0.0-alpha.3_local_missing_comma`, { encoding: 'utf8' }))();
 
-      const expectedPrintedErrorMessage = 'Error occurred while instantiating the model. Details follow: Invalid syntax in "Job": Unexpected string in JSON at position 1005';
+      const expectedPrintedErrorMessage = 'Error occurred while instantiating the model. Details follow: Invalid syntax in "Job": Unexpected string in JSON at position 1142';
 
       await rewiredApi.testPlans(jobFileContents);
 
@@ -188,7 +187,7 @@ describe('apiDecoratingAdapter', () => {
     it('- on - invalid job based on purpleteam schema - should print useful error message', async (flags) => {
       // Lots of checking around the validation on the server side will be required.
       const { context: { rewiredApi, critStub } } = flags;
-      const jobFileContents = await (async () => readFileAsync(`${process.cwd()}/testResources/jobs/job_1.0.0-alpha.3_local_missing_type_of_appScanner`, { encoding: 'utf8' }))();
+      const jobFileContents = await (async () => readFileAsync(`${process.cwd()}/testResources/jobs/job_2.0.0-alpha.3_local_missing_type_of_appScanner`, { encoding: 'utf8' }))();
 
       const expectedResponseBodyMessage = '';// Doesn't matter what this is, we don't check it.
       /* eslint-disable no-useless-escape */
