+++ testResources/jobs/job_2.0.0-alpha.3_local_missing_type_of_appScanner
@@ -1,0 +1,123 @@
+{
+  "data": {
+    "type": "BrowserApp",
+    "attributes": {
+      "version": "2.0.0-alpha.3",
+      "sutAuthentication": {
+        "sitesTreeSutAuthenticationPopulationStrategy": "FormStandard",
+        "emissaryAuthenticationStrategy": "FormStandard",
+        "route": "/login",
+        "usernameFieldLocater": "userName",
+        "passwordFieldLocater": "password",
+        "submit": "btn btn-danger",
+        "expectedPageSourceSuccess": "Log Out"
+      },
+      "sutIp": "pt-sut-cont",
+      "sutPort": 4000,
+      "sutProtocol": "http",
+      "browser": "chrome",
+      "loggedInIndicator": "<p>Found. Redirecting to <a href=\"\/dashboard\">\/dashboard<\/a><\/p>"
+    },
+    "relationships": {
+      "data": [{
+        "type": "tlsScanner",
+        "id": "NA"
+      },
+      {
+        "type": "appScanner",
+        "id": "lowPrivUser"
+      },
+      {
+        "type": "appScanner",
+        "id": "adminUser"
+      }]
+    }
+  },
+  "included": [
+    {
+      "type": "tlsScanner",
+      "id": "NA",
+      "attributes": {
+        "tlsScannerSeverity": "LOW",
+        "alertThreshold": 3
+      }
+    },
+    {
+      "id": "lowPrivUser",
+      "attributes": {
+        "sitesTreePopulationStrategy": "WebDriverStandard",
+        "spiderStrategy": "BrowserAppStandard",
+        "scannersStrategy": "BrowserAppStandard",
+        "scanningStrategy": "Standard",
+        "postScanningStrategy": "Standard",
+        "reportingStrategy": "Standard",
+        "username": "user1",
+        "password": "User1_123",
+        "aScannerAttackStrength": "HIGH",
+        "aScannerAlertThreshold": "LOW",
+        "alertThreshold": 12
+      },
+      "relationships": {
+        "data": [{
+          "type": "route",
+          "id": "/profile"
+        }]
+      }
+    },
+    {
+      "type": "appScanner",
+      "id": "adminUser",
+      "attributes": {
+        "sitesTreePopulationStrategy": "WebDriverStandard",
+        "spiderStrategy": "BrowserAppStandard",
+        "scannersStrategy": "BrowserAppStandard",
+        "scanningStrategy": "Standard",
+        "postScanningStrategy": "Standard",
+        "reportingStrategy": "Standard",
+        "username": "admin",
+        "password": "Admin_123"
+      },
+      "relationships": {
+        "data": [{
+          "type": "route",
+          "id": "/memos"
+        },
+        {
+          "type": "route",
+          "id": "/profile"
+        }]
+      }
+    },
+    {
+      "type": "route",
+      "id": "/profile",
+      "attributes": {
+        "attackFields": [
+          {"name": "firstName", "value": "PurpleJohn", "visible": true},
+          {"name": "lastName", "value": "PurpleDoe", "visible": true},
+          {"name": "ssn", "value": "PurpleSSN", "visible": true},
+          {"name": "dob", "value": "12235678", "visible": true},
+          {"name": "bankAcc", "value": "PurpleBankAcc", "visible": true},
+          {"name": "bankRouting", "value": "0198212#", "visible": true},
+          {"name": "address", "value": "PurpleAddress", "visible": true},
+          {"name": "website", "value": "https://purpleteam-labs.com", "visible": true},
+          {"name": "_csrf", "value": ""},
+          {"name": "submit", "value": ""}
+        ],
+        "method": "POST",
+        "submit": "submit"
+      }
+    },
+    {
+      "type": "route",
+      "id": "/memos",
+      "attributes": {
+        "attackFields": [
+          {"name": "memo", "value": "PurpleMemo", "visible": true}
+        ],
+        "method": "POST",
+        "submit": "btn btn-primary"
+      }
+    }
+  ]
+}
